feat(.github/workflows): Update workflow to delete tags greater than v12.0
In this update, the Github Actions workflow configuration has been radically altered. The old workflow, that used to trigger CI builds for pushes and pull requests to the main branch, is no longer active.

Instead, a new workflow has been put in place which triggers whenever a new tag in the format 'v*' is created in the repository. The new job : 'delete-tag', checks if the tag version is greater than 'v12.0'. If true, the job automatically deletes the newly created tag both locally and remotely. This aims to ensure that tags don't exceed the set version of 'v12.0'.🗑️🔖🧹