feat: Add explanatory documentation for Tag Deletion Demonstration 👨‍💻
This commit adds detailed explanation about Tag Deletion Demonstration project wherein a mechanism to automatically delete a Git tag, if pushed by mistake, is shown. It explains the use of GitHub Actions in monitoring tag creation in the repository. When a new tag is created, a workflow triggers and checks if the tag name is greater than "v12.0". If so, the tag is deleted both locally and from the remote repository. The documentation includes broader details about the working logic, sample workflow code, and mention of Protected Tags as a secure method for the same task. It also provides guidance to navigate towards "Tags" section in the repository settings, along with a link to the GitHub documentation on tag protection rules.